requirements.txt
absl-py==0.13.0
cachetools==4.2.2
charset-normalizer==2.0.3
cloudpickle==1.6.0
cycler==0.10.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.40.0
hyperopt==0.2.5
idna==3.2
isodate==0.6.0
jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
markdown==3.3.4
markupsafe==2.0.1
matplotlib==3.4.2
memory-profiler==0.58.0
numpy==1.21.0
oauthlib==3.1.1
pandas==1.3.0
profiler==0.1.0
protobuf==3.18.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynvml==11.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-louvain==0.15
pytz==2021.1
rdflib==6.0.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
setuptools==44.1.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
texttable==1.6.4
threadpoolctl==2.1.0
torch-cluster==1.5.9
torch-geometric==1.7.2
torch-scatter==2.0.7
torch-sparse==0.6.10
torch-spline-conv==1.2.1
torchstat==0.0.7
tqdm==4.61.2
urllib3==1.26.6
werkzeug==2.0.1